chatlog data structure and fetching from server
finalized the data structure that will contain 100 chat messages in server side memory.  It is a dictionary of chat channels each containing a dictionary of timestamps containing an array storing user name in index 0 and the chat message in index 1.

Also refined the user interface for the chatbox and tested fetching  and displaying the chatlog from the server.